drop and recreate indices for df

+++ custom/icds_reports/utils/aggregation_helpers/distributed/base.py
@@ -70,14 +70,17 @@ class StateBasedAggregationDistributedHelper(BaseICDSAggregationDistributedHelpe
     months_required = None
 
     def aggregate(self, cursor):
-        delete_query, delete_params = self.delete_previous_run_query()
+        if self.delete_previous_run_query():
+            delete_query, delete_params = self.delete_previous_run_query()
         agg_query, agg_params = self.aggregation_query()
 
-        logger.info(f'Deleting old data for {self.helper_key} month {self.month} and state {self.state_id}')
-        if self.months_required:
+        if self.months_required and self.delete_old_data_query():
+            logger.info(f'Deleting old data for {self.helper_key} month {self.month} and state {self.state_id}')
+
             cursor.execute(*self.delete_old_data_query())
-        logger.info(f'Deleting for {self.helper_key} month {self.month} and state {self.state_id}')
-        cursor.execute(delete_query, delete_params)
+        if self.delete_previous_run_query():
+            logger.info(f'Deleting for {self.helper_key} month {self.month} and state {self.state_id}')
+            cursor.execute(delete_query, delete_params)
         logger.info(f'Starting aggregation for {self.helper_key} month {self.month} and state {self.state_id}')
         cursor.execute(agg_query, agg_params)
         logger.info(f'Finished aggregation for {self.helper_key} month {self.month} and state {self.state_id}')
@@ -97,6 +100,23 @@ class StateBasedAggregationDistributedHelper(BaseICDSAggregationDistributedHelpe
             f'DELETE FROM "{self.aggregate_parent_table}" WHERE month < %(month)s AND state_id = %(state)s',
             {'month': month_formatter(month), 'state': self.state_id}
         )
+
+    def delete_queries(self):
+        queries = [(
+            f'DELETE FROM "{self.aggregate_parent_table}" WHERE month=%(month)s',
+            {'month': month_formatter(self.month)}
+        )]
+
+        if self.months_required is None:
+            return queries
+
+        month = self.month - relativedelta(months=self.months_required)
+        queries.append((
+            f'DELETE FROM "{self.aggregate_parent_table}" WHERE month < %(month)s',
+            {'month': month_formatter(month)}
+        ))
+
+        return queries
 
 
 class StateBasedAggregationPartitionedHelper(BaseICDSAggregationDistributedHelper):
